requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
backcall==0.2.0
catalogue==2.0.4
certifi==2021.5.30
charset-normalizer==2.0.2
click==7.1.2
decorator==5.0.9
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.66.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.2
idna==3.2
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
orjson==3.6.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
promise==2.3
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pydantic==1.8.2
Pygments==2.9.0
python-dotenv==0.18.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
Rx==1.6.1
six==1.16.0
srsly==2.4.1
starlette==0.14.2
traitlets==5.0.5
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.15.3
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
